refactor for source types
A beginning of the source types refactoring into their own files/classes
